BUGFIX: Avoid overflowing the measurement fifos
Before inserting a measurement into a full fifo, we now discard the
oldest item.

In order to detect when the fifo is full, we now allocate an array which
is one cell larger than the requested fifo size. The field FifoSize in
CHANNELPARAMETER, which is part of the visible API, is the usable size
of the fifo, whereas the field Size in FIFO is the allocated size.