Support libgpiod v2
The API version (v1 or v2) is automatically detected by TrmsLin.c.

Note that the apt package "libgpiod2" ships libgpiod v1.6.3. The suffix
"2" in the package name comes from the SONAME, which is libgpiod.so.2.
Debian (and Raspberry Pi OS) does not yet ship libgpiod v2, which has
the SONAME libgpiod.so.3.